+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add manual
+- Support Jackson subtypes as discriminated unions ([#50](https://github.com/EvidentSolutions/apina/pull/50))
 - Change `ANGULAR2` platform name to `ANGULAR`.
 - Support other nullability annotations in addition to JetBrains' annotations.